Move build.rs-specific functions out of cef-ui-util
